+++ content/footer.php
@@ -1,5 +1,5 @@
-		</div>
-		<footer>
+		</main>
+		<footer class="footer">
 		
 		</footer>
 		<?php getThemeScript('script.js'); ?>