+++ includes/class-post.php
@@ -7,20 +7,40 @@
  */
 class Post {
 	/**
+	 * The currently queried post's slug.
+	 * @since 2.2.3[a]
+	 *
+	 * @access private
+	 * @var string
+	 */
+	private $slug;
+	
+	/**
+	 * Class constructor. Sets the default queried post slug.
+	 * @since 2.2.3[a]
+	 *
+	 * @access public
+	 * @param string $slug (optional; default: '')
+	 * @return null
+	 */
+	public function __construct($slug = '') {
+		$this->slug = $slug;
+	}
+	
+	/**
 	 * Fetch a post's id.
 	 * @since 2.2.0[a]
 	 *
 	 * @access public
-	 * @param string $slug
 	 * @param bool $echo (optional; default: true)
 	 * @return null|int (null on $echo == true; int on $echo == false)
 	 */
-	public function getPostId($slug, $echo = true) {
+	public function getPostId($echo = true) {
 		// Extend the Query class
 		global $rs_query;
 		
 		// Fetch the post's id from the database
-		$id = (int)$rs_query->selectField('posts', 'id', array('slug'=>$slug));
+		$id = (int)$rs_query->selectField('posts', 'id', array('slug'=>$this->slug));
 		
 		if($echo)
 			echo $id;
@@ -33,16 +53,15 @@ class Post {
 	 * @since 2.2.0[a]
 	 *
 	 * @access public
-	 * @param string $slug
 	 * @param bool $echo (optional; default: true)
 	 * @return null|string (null on $echo == true; string on $echo == false)
 	 */
-	public function getPostTitle($slug, $echo = true) {
+	public function getPostTitle($echo = true) {
 		// Extend the Query class
         global $rs_query;
 		
 		// Fetch the post's title from the database
-        $title = $rs_query->selectField('posts', 'title', array('slug'=>$slug));
+        $title = $rs_query->selectField('posts', 'title', array('slug'=>$this->slug));
 		
         if($echo)
             echo $title;
@@ -55,16 +74,15 @@ class Post {
 	 * @since 2.2.0[a]
 	 *
 	 * @access public
-	 * @param string $slug
 	 * @param bool $echo (optional; default: true)
 	 * @return null|string (null on $echo == true; string on $echo == false)
 	 */
-	public function getPostAuthor($slug, $echo = true) {
+	public function getPostAuthor($echo = true) {
 		// Extend the Query class
         global $rs_query;
 		
 		// Fetch the post's author from the database
-        $author = $rs_query->selectField('posts', 'author', array('slug'=>$slug));
+        $author = $rs_query->selectField('posts', 'author', array('slug'=>$this->slug));
 		
 		// Fetch the author's username from the database
 		$username = $rs_query->selectField('users', 'username', array('id'=>$author));
@@ -80,16 +98,15 @@ class Post {
 	 * @since 2.2.0[a]
 	 *
 	 * @access public
-	 * @param string $slug
 	 * @param bool $echo (optional; default: true)
 	 * @return null|string (null on $echo == true; string on $echo == false)
 	 */
-	public function getPostDate($slug, $echo = true) {
+	public function getPostDate($echo = true) {
 		// Extend the Query class
         global $rs_query;
 		
 		// Fetch the post's title from the database
-        $date = $rs_query->selectField('posts', 'date', array('slug'=>$slug));
+        $date = $rs_query->selectField('posts', 'date', array('slug'=>$this->slug));
 		
         if($echo)
             echo formatDate($date, 'd M Y @ g:i A');
@@ -102,16 +119,15 @@ class Post {
 	 * @since 2.2.0[a]
 	 *
 	 * @access public
-	 * @param string $slug
 	 * @param bool $echo (optional; default: true)
 	 * @return null|string (null on $echo == true; string on $echo == false)
 	 */
-	public function getPostModDate($slug, $echo = true) {
+	public function getPostModDate($echo = true) {
 		// Extend the Query class
         global $rs_query;
 		
 		// Fetch the post's title from the database
-        $modified = $rs_query->selectField('posts', 'modified', array('slug'=>$slug));
+        $modified = $rs_query->selectField('posts', 'modified', array('slug'=>$this->slug));
 		
         if($echo)
             echo formatDate($modified, 'd M Y @ g:i A');
@@ -124,16 +140,15 @@ class Post {
 	 * @since 2.2.0[a]
 	 *
 	 * @access public
-	 * @param string $slug
 	 * @param bool $echo (optional; default: true)
 	 * @return null|string (null on $echo == true; string on $echo == false)
 	 */
-	public function getPostContent($slug, $echo = true) {
+	public function getPostContent($echo = true) {
 		// Extend the Query class
         global $rs_query;
 		
 		// Fetch the post's content from the database
-        $content = $rs_query->selectField('posts', 'content', array('slug'=>$slug));
+        $content = $rs_query->selectField('posts', 'content', array('slug'=>$this->slug));
 		
 		// Filter out any HTML or JavaScript comments
         //$filtered_content = $this->removeComments($post['content']);
@@ -149,16 +164,15 @@ class Post {
 	 * @since 2.2.0[a]
 	 *
 	 * @access public
-	 * @param string $slug
 	 * @param bool $echo (optional; default: true)
 	 * @return null|string (null on $echo == true; string on $echo == false)
 	 */
-	public function getPostStatus($slug, $echo = true) {
+	public function getPostStatus($echo = true) {
 		// Extend the Query class
         global $rs_query;
 		
 		// Fetch the post's status from the database
-        $status = $rs_query->selectField('posts', 'status', array('slug'=>$slug));
+        $status = $rs_query->selectField('posts', 'status', array('slug'=>$this->slug));
 		
         if($echo)
             echo $status;
@@ -171,7 +185,7 @@ class Post {
 	 * @since 2.2.0[a]
 	 *
 	 * @access public
-	 * @param string $slug
+	 * @param int $id
 	 * @param bool $echo (optional; default: true)
 	 * @return null|string (null on $echo == true; string on $echo == false)
 	 */
@@ -193,16 +207,15 @@ class Post {
 	 * @since 2.2.0[a]
 	 *
 	 * @access public
-	 * @param string $slug
 	 * @param bool $echo (optional; default: true)
 	 * @return null|string (null on $echo == true; string on $echo == false)
 	 */
-	public function getPostParent($slug, $echo = true) {
+	public function getPostParent($echo = true) {
 		// Extend the Query class
         global $rs_query;
 		
 		// Fetch the post's parent from the database
-        $parent = $rs_query->selectField('posts', 'parent', array('slug'=>$slug));
+        $parent = $rs_query->selectField('posts', 'parent', array('slug'=>$this->slug));
 		
         if($echo)
             echo $parent;
@@ -215,16 +228,15 @@ class Post {
 	 * @since 2.2.0[a]
 	 *
 	 * @access public
-	 * @param string $slug
 	 * @param bool $echo (optional; default: true)
 	 * @return null|string (null on $echo == true; string on $echo == false)
 	 */
-	public function getPostType($slug, $echo = true) {
+	public function getPostType($echo = true) {
 		// Extend the Query class
 		global $rs_query;
 		
 		// Fetch the post's type from the database
-		$type = $rs_query->selectField('posts', 'type', array('slug'=>$slug));
+		$type = $rs_query->selectField('posts', 'type', array('slug'=>$this->slug));
 		
 		if($echo)
 			echo $type;
@@ -237,20 +249,56 @@ class Post {
 	 * @since 2.2.0[a]
 	 *
 	 * @access public
-	 * @param string $slug
 	 * @param bool $echo (optional; default: true)
 	 * @return null|string (null on $echo == true; string on $echo == false)
 	 */
-	public function getPostFeatImage($slug, $echo = true) {
+	public function getPostFeatImage($echo = true) {
 		// Extend the Query class
 		global $rs_query;
 		
 		// Fetch the featured image's id from the database
-		$feat_image = (int)$rs_query->selectField('postmeta', 'value', array('post'=>$this->getPostId($slug, false), '_key'=>'feat_image'));
+		$feat_image = (int)$rs_query->selectField('postmeta', 'value', array('post'=>$this->getPostId(false), '_key'=>'feat_image'));
 		
         if($echo)
             echo getMedia($feat_image, 'featured-image');
         else
             return getMedia($feat_image, 'featured-image');
+    }
+	
+	/**
+	 * Fetch a post's metadata.
+	 * @since 2.2.3[a]
+	 *
+	 * @access public
+	 * @param string $key
+	 * @param bool $echo (optional; default: true)
+	 * @return null|string (null on $echo == true; string on $echo == false)
+	 */
+	public function getPostMeta($key, $echo = true) {
+		// Extend the Query class
+        global $rs_query;
+		
+		// Fetch the post's metadata from the database
+        $meta = $rs_query->selectField('postmeta', 'value', array('post'=>$this->getPostId(false), '_key'=>$key));
+
+        if($echo)
+            echo $meta;
+        else
+            return $meta;
+    }
+	
+	/**
+	 * Fetch a post's full URL.
+	 * @since 2.2.3[a]
+	 *
+	 * @access public
+	 * @param bool $echo (optional; default: true)
+	 * @return null|string (null on $echo == true; string on $echo == false)
+	 */
+	public function getPostUrl($echo = true) {
+        if($echo)
+            echo getSetting('site_url', false).getPermalink($this->getPostParent(false), $this->slug);
+        else
+            return getSetting('site_url', false).getPermalink($this->getPostParent(false), $this->slug);
     }
 }